+++ src/server/api/plants.js
@@ -110,71 +110,63 @@ router.delete("/account/favorite/:plantId", async (req, res, next) => {
     }
 });
 
-// validates if notes exist and assigned to user
-const validateNotes = ( user, notes ) => {
-    if (!notes) {
-        throw new ServerError(404, "Notes not found.");
-    } if (notes.userId !== user.id) {
-        throw new ServerError(403, "No notes for this user.")
-    }
-};
 
-// sends all notes
-router.get("/account/mynotes", async (req, res, next) => {
-    try {
-        // validates user is logged in
-        // finds many notes
-        const notes = await prisma.note.findMany({
-            where: { userId: res.locals.user.id }
-        });
-        validateNotes(res.locals.user, notes);
-        res.json(notes)
-    } catch (err) {
-        next (err);
-    }
-});
+
+// // sends all notes
+// router.get("/account/mynotes", async (req, res, next) => {
+//     try {
+      
+//         const notes = await prisma.note.findMany({
+//             where: { userId: res.locals.user.id }
+//         });
+//         validateNotes(res.locals.user, notes);
+//         res.json(notes)
+//     } catch (err) {
+//         next (err);
+//     }
+// });
 
 // creates a new note in the account page
-router.post("/account/mynotes", async (req, res, next) => {
-    try {
-        // validates user is logged in
-        // requires an id and user info
-        const { note } = req.body;
-        if (!description) {
-            throw new ServerError(400, "Description required.");
-        }
-        const note = await prisma.note.create({
-            data: {
-                id,
-                description, 
-                user: { connect: { id: res.locals.user.id }},
-            },
-        });
-        res.json(note);
-    } catch (err) {
-        next (err);
-    }
-});
+// router.post("/account/mynotes", async (req, res, next) => {
+//     try {
+//         // validates user is logged in
+//         // requires an id and user info
+//         const { note } = req.body;
+//         if (!description) {
+//             throw new ServerError(400, "Description required.");
+//         }
+//         const newNote = await prisma.note.create({
+//             data: {
+//                 id,
+//                 description, 
+//                 user: { connect: { id: res.locals.user.id }},
+//             },
+//         });
+//         res.json(note);
+//     } catch (err) {
+//         next (err);
+//     }
+// });
 
 // puts an edit on a note in the account page
-router.put("/account/mynotes/:id", async (req, res, next) => {
-    try {
-        const id = +req.params.id;
-        const { note } = req.body;
-        // validates user is logged in
-        // requires an id and user info
-        const findNote = await prisma.note.findUnique({ where: { id }});
-        validateNotes(res.locals.user, note);
+// router.put("/account/mynotes/:id", async (req, res, next) => {
+//     try {
+//         const id = +req.params.id;
+//         const { note } = req.body;
+//         // validates user is logged in
+//         // requires an id and user info
+//         const findNote = await prisma.note.findUnique({ where: { id }});
+//         validateNotes(res.locals.user, note);
 
-        const updatedNote = await prisma.note.update({
-            where: { id }, 
-            data: { note },
-        });
-        res.json(updatedNote);
-    } catch (err) {
-        next (err);
-    }
-});
+//         const updatedNote = await prisma.note.update({
+//             where: { id }, 
+//             data: { note },
+//         });
+//         res.json(updatedNote);
+//     } catch (err) {
+//         next (err);
+//     }
+// });
 
 // deletes a note in account page
 router.delete("/account/mynotes/:id", async (req, res, next) => {
